+++ datawarga/kependudukan/templates/base.html
@@ -90,6 +90,12 @@
                                     <i class="bi bi-filetype-csv"></i> Import Data Warga
                                 </a>
                             </li>
+                            <li class="nav-item">
+                                <a class="nav-link" href="{% url 'kependudukan:utilAssignWargaRumah' %}">
+                                    <span class="align-text-bottom"></span>
+                                    <i class="bi bi-arrow-left-right"></i> Pindahkan Warga
+                                </a>
+                            </li>
                             <li class="nav-item d-sm-none">
                                 <a class="nav-link" href="{% url 'logout' %}">
                                     <span class="align-text-bottom"></span>
