+++ datawarga/kependudukan/templates/form_kompleks_detail.html
@@ -66,7 +66,9 @@
         <div class="py-2">
             <h5>Data Warga Penghuni</h5>
             <p>Berikut ini adalah daftar warga penghuni blok {{ data.blok }} / No. {{ data.nomor }}</p>
-            <p><a href="{% url 'kependudukan:formWargaRumah' data.id %}" class="btn btn-primary btn-sm"><i class="bi bi-plus"></i></a></p>
+            <p>
+                <a href="{% url 'kependudukan:formWargaRumah' data.id %}" class="btn btn-primary btn-sm"><i class="bi bi-plus"></i> Tambah Warga Baru</a> 
+            </p>
             <div id="tabelwarga"></div>
         </div>
     </div>
